// src/models/CodeBlock.ts
export interface CodeBlock {
  object: "block";
  type: "code";
  code: {
    caption: [];
    rich_text: {
      type: "text";
      text: {
        content: string;
      };
    }[];
    language:
      | "abap"
      | "arduino"
      | "bash"
      | "basic"
      | "c"
      | "clojure"
      | "coffeescript"
      | "c++"
      | "c#"
      | "css"
      | "dart"
      | "diff"
      | "docker"
      | "elixir"
      | "elm"
      | "erlang"
      | "flow"
      | "fortran"
      | "f#"
      | "gherkin"
      | "glsl"
      | "go"
      | "graphql"
      | "groovy"
      | "haskell"
      | "html"
      | "java"
      | "javascript"
      | "json"
      | "julia"
      | "kotlin"
      | "latex"
      | "less"
      | "lisp"
      | "livescript"
      | "lua"
      | "makefile"
      | "markdown"
      | "markup"
      | "matlab"
      | "mermaid"
      | "nix"
      | "objective|c"
      | "ocaml"
      | "pascal"
      | "perl"
      | "php"
      | "plain text"
      | "powershell"
      | "prolog"
      | "protobuf"
      | "python"
      | "r"
      | "reason"
      | "ruby"
      | "rust"
      | "sass"
      | "scala"
      | "scheme"
      | "scss"
      | "shell"
      | "sql"
      | "swift"
      | "typescript"
      | "vb.net"
      | "verilog"
      | "vhdl"
      | "visual basic"
      | "webassembly"
      | "xml"
      | "yaml"
      | "java/c/c++/c#";
  };
}